move user stats as well
