+++ airflow/www/security.py
@@ -37,6 +37,7 @@ from airflow.auth.managers.fab.views.user_edit import (
     CustomResetPasswordView,
     CustomUserInfoEditView,
 )
+from airflow.auth.managers.fab.views.user_stats import CustomUserStatsChartView
 from airflow.exceptions import AirflowException, RemovedInAirflow3Warning
 from airflow.models import DagBag, DagModel
 from airflow.security import permissions
@@ -47,7 +48,6 @@ from airflow.www.fab_security.sqla.manager import SecurityManager
 from airflow.www.fab_security.views import (
     ActionModelView,
     CustomRoleModelView,
-    CustomUserStatsChartView,
     PermissionPairModelView,
     ResourceModelView,
 )
